made some improvements and cleaned up code
