+++ examples/toy_model/run_samplers.py
@@ -112,8 +112,8 @@ def main():
             #"run": True,
             "run": False,
             "n_chains": 1,
-            "n_steps": 1000,
-            "save_freq": 10,
+            "n_steps": 10000,
+            "save_freq": 100,
             "use_sigma_dist": True, # Example of sampler-specific parameter
         },
          "octet_sampler": { # Example configuration for OctamerSampler
